try send 0.4s and just send one symbol. (the symbol rate 10).
Let's suppose we want send a 2 in C0 and zero in other subchannels. we send it in 400 samples for 0.4 s just for 1 symbol. Then we get:
 | C0 | C1 | C2 | C3 | C4 | C5 | C6 | C7 | C8 | C9 |
 | 2  | 0  | 0  | 0  | 0  | 0  | 0  | 0  | 0  | 0  | last 0.1s

From the the generated diagram, we can see lost samples in time domain is nearly 0. The is can help to separate the  symbol when Receive signal.