Add 0.005 s guard period for one symbol (One symbol length is 0.1s). for LTE: CP is 4.69 µs while symbol length in LTE set to 66.7 µs.
If not long distance communication, the ISI will not so big, the very small guard period (even no CP) is OK. if for very long distance.
for example: For LTE, the standard length of the cyclic prefix has been chosen to be 4.69 µs. This enables the system to accommodate path variations of up to 1.4 km. With the symbol length in LTE set to 66.7 µs.